Restructure the write code
Move the multiblock code into a separate file and introduce the usage of
record_functions_st for some write functions.
